VDB-6118 fixed Windows compilation warnings/errors

+++ interfaces/kapp/vdbapp.h
@@ -189,9 +189,9 @@ rc_t CC NextLogLevelCommon ( const char * level_parameter );
             void setRc( rc_t p_rc ) { m_rc = p_rc; }
 
             // recommended exit code for main() based on reported rc
-            int getExitCode() const { return m_rc == 0 ? 0 : IF_EXITCODE( m_rc, 3 ); }
+            int getExitCode() const { return ( m_rc == 0 ) ? 0 : (int)(IF_EXITCODE( m_rc, 3 )); }
 
-            int getArgC() const { return m_argc; }
+            int getArgC() const { return (int)m_argc; }
             char** getArgV() { return m_argv; }
             const char** getArgV() const { return (const char**)m_argv; }
 
@@ -200,7 +200,7 @@ rc_t CC NextLogLevelCommon ( const char * level_parameter );
         private:
             bool IsStandardOption( unsigned int index, bool & skipNext ) const;
 
-            int m_argc;
+            unsigned int m_argc;
             char** m_argv;
             bool m_argvOwned; // true if args have been rewritten
             rc_t m_rc = 0;
